Stairs

Calculates the staircase-ness measure defined in Norstrom et al., Evolutionary Bioinformatics online, 8:261 (2012) doi:10.4137/EBO.S9738 .

stairs(tree)

Arguments

tree

a tree of class phylo or phylo4. The tree should be binary and rooted; if not it will be coerced into a binary rooted tree using multi2di, if possible.

Value

Two numbers corresponding to the two staircase-ness measures for a tree. These are two related measures:

1: the proportion of subtrees that are imbalanced (i.e. subtrees where the left child has more tip descendants than the right child, or vice versa)

2: the average of all the min(l,r)/max(l,r) values of each subtree, where l and r are the number of tips in the left and right children of a subtree.

Author

Michelle Kendall michelle.louise.kendall@gmail.com

Examples

## Find the staircase-ness measures in a random tree with 20 tips:
stairs(rtree(20))
#> [1] 0.6315789 0.6334757